requirements.txt
librosa==0.10.1
Pillow==9.4.0
numpy==1.24.2
